requirements.txt
aioredis==1.3.1
amqp==5.0.6
asgiref==3.3.4
async-timeout==3.0.1
attrs==21.2.0
autobahn==21.3.1
Automat==20.2.0
billiard==3.6.4.0
celery==5.0.5
certifi==2020.12.5
cffi==1.14.5
channels==3.0.3
channels-redis==3.2.0
chardet==4.0.0
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
constantly==15.1.0
cryptography==3.4.7
daphne==3.0.2
Django==3.2.2
django-cors-headers==3.7.0
django-filter==2.4.0
django-model-utils==4.1.1
django-notifications-hq==1.6.0
django-notifications-rest==0.0.1
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
fcm-django==0.3.10
hiredis==2.0.0
hyperlink==21.0.0
idna==2.10
incremental==21.3.0
jsonfield==3.1.0
kombu==5.1.0
Markdown==3.3.4
msgpack==1.0.2
mysqlclient==2.0.3
prompt-toolkit==3.0.18
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyfcm==1.4.9
PyJWT==1.7.1
pyOpenSSL==20.0.1
pytz==2021.1
requests==2.25.1
service-identity==21.1.0
six==1.16.0
sqlparse==0.4.1
swapper==1.1.2.post1
Twisted==21.2.0
twisted-iocpsupport==1.0.1
txaio==21.2.1
urllib3==1.26.5
vine==5.0.0
wcwidth==0.2.5
zope.interface==5.4.0